pyproject.toml
[project]
name = "python-playwright-ai"
version = "0.1.0"
description = "UI test automation framework using Playwright, Pytest, and Allure"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "allure-pytest>=2.15.3",
    "faker>=40.12.0",
    "playwright>=1.58.0",
    "pydantic-settings>=2.5.2",
]

[dependency-groups]
dev = [
    "black>=24.4.2",
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
pw_browser = "chromium"
pw_headless = "false"
pw_base_url = "https://demoqa.com"
pw_tracing = "retain-on-failure"
pw_screenshot = "only-on-failure"
pw_video = "retain-on-failure"
pw_slowmo = "0"
pw_viewport_width = "1920"
pw_viewport_height = "1080"
pw_timeout = "30000"
pw_output_dir = "test-results"
pw_full_page_screenshot = "false"
pw_fullscreen = "true"

markers = [
    "smoke: Quick, critical path tests",
    "regression: Comprehensive suite"
]
addopts = [
    "-v",
    "--tb=short",
    "--alluredir=allure-results",
    "--clean-alluredir",
]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py314']

[tool.ruff]
line-length = 88
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
disallow_untyped_defs = true
